huddle-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
deepgram-sdk==3.2.7
google-generativeai==0.3.2
PyMuPDF==1.23.8
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
aiofiles==23.2.1
websockets==12.0